% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_mfbavart.R
\name{soft_mfbavart}
\alias{soft_mfbavart}
\title{Mixed Frequency Soft Bayesian Additive Vector Autoregression Trees (Huber et al. 2023)}
\usage{
soft_mfbavart(
  data,
  itr,
  p = 5,
  fhorz = 0,
  cons = FALSE,
  exact = FALSE,
  sv = FALSE,
  var.thrsh = 10,
  max.count.var = 10,
  cgm.level = 0.95,
  cgm.exp = 2,
  sd.mu = 2,
  num.trees = 250,
  nburn = 1000,
  nsave = 1000,
  thinfac = 1,
  quiet = FALSE,
  SB_group = NULL,
  SB_alpha = 1,
  SB_beta = 2,
  SB_gamma = 0.95,
  SB_k = 2,
  SB_sigma_hat = NULL,
  SB_shape = 1,
  SB_width = 0.1,
  SB_alpha_scale = NULL,
  SB_alpha_shape_1 = 0.5,
  SB_alpha_shape_2 = 1,
  SB_tau_rate = 10,
  SB_num_tree_prob = NULL,
  SB_temperature = 1,
  SB_weights = NULL,
  SB_normalize_Y = TRUE
)
}
\arguments{
\item{data}{list of data frames for each variable.}

\item{p}{number of lags}

\item{fhorz}{number of forecast horizons}

\item{ite}{option for constructing the loading matrix. Can be "m", "lvl", or "grw".}
}
\value{
The following objects are returned:
\item{Y}{Array of in-sample posterior draws. First dimension is number of MCMC iterations, second dimension is number of training time periods, third dimension is number of variables (length of input data list).}
\item{fcst}{Array of out-of-sample posterior draws (including noise form error term). First dimension is number of MCMC iterations, second dimension is number of horecast horizons, third dimension is number of variables.}
\item{Yq}{Array of out-of-sample quarterly aggregates. First dimension is number of MCMC iterations, second dimension is number of horecast horizons, third dimension is number of variables.}
}
\description{
Mixed Frequency Soft Bayesian Additive Vector Autoregression Trees (Huber et al. 2023).
}
\examples{
# -----------------------------------------------------------------------------
# load dataset
library(alfred) # load dataset from FRED (Fed St. Louis)

variables <- c("CPIAUCSL","UNRATE","GDPC1")
out <- lapply(variables, get_alfred_series,observation_start = "1980-01-01",observation_end = "2020-12-01",realtime_start = "2020-12-01",realtime_end = "2020-12-01")
alfred_to_ts <- function(x, freq){
  ts(x[,3],start=c(1980,1),frequency=freq)
}
mf_list <- mapply(alfred_to_ts, x = out, freq = c(12, 12, 4))
names(mf_list) <- variables
log_diff <- function(x) {
  freq <- frequency(x)
  100 * freq * diff(log(x))}
mf_list[c("CPIAUCSL", "GDPC1")] <- lapply(mf_list[c("CPIAUCSL", "GDPC1")], log_diff)
mf_list <- mapply(window, x = mf_list, start = list(c(1980, 4), c(1980, 4), c(1980, 2)))
data <- mf_list

est_obj <- mfbavart(data,itr="grw",fhorz=2,prior.sig=c(200,0.75))

Yq <- est_obj$Yq
GDPC1_post <- t(apply(Yq,c(2,3),quantile,probs=c(0.16,0.5,0.84),na.rm=TRUE)[,,"GDPC1"])
ts.plot(GDPC1_post)
}
